Delete Prompt now works for Todo component

+++ client/src/components/Todo.jsx
@@ -1,53 +1,73 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 
-import { Button } from '@material-ui/core';
 import ClearIcon from '@mui/icons-material/Clear';
 
-function Todo({ todoList, setTodoList }) {
+import DeletePromptModal from './DeletePromptModal.jsx'
 
-  const mapTodoList = () =>
+function Todo({ todoList, setTodoList }) {
+  const [deletePromptChecked, setDeletePromptChecked] = useState(false);
+
+  const deleteTodoTask = (indexNum) => {
+    const todoSplice = () => {todoList.splice(indexNum, 1)};
+
+    todoSplice();
+    setTodoList([...todoList]);
+  };
+
+  const handleDeletePromptCheck = (index) => {
+    if (deletePromptChecked === false) {
+      return (
+        <DeletePromptModal
+            index={index}
+            deleteTodoTask={deleteTodoTask}
+            deletePromptChecked={deletePromptChecked}
+            setDeletePromptChecked={setDeletePromptChecked}
+            handleDeletePromptCheck={handleDeletePromptCheck}
+          />
+      )
+    } else if (deletePromptChecked === true) {
+      return (
+        <ClearIcon
+          className="task-close-btn"
+          onClick={ () => {deleteTodoTask(index)}}
+        />
+      )
+    }
+  };
+
+  const mapTodoList = () =>(
 
     todoList.map((todoMsg, index) => (
       <div
         key={index}
         className="task-container"
       >
-        <div
-          className="task-close-btn-cont"
-
-        >
-          <ClearIcon
-          className="task-close-btn"
-          onClick={() => {deleteTodoTask(index)}}
-          />
+        <div className="task-close-btn-cont">
+          { handleDeletePromptCheck(index) }
         </div>
+
         <div className="task-cont-msg">
         {todoMsg}
         </div>
-      </div>
-    ));
 
-  const deleteTodoTask = (indexNum) => {
-    const todoSplice = () => {todoList.splice(indexNum, 1)};
-    todoSplice();
-    setTodoList([...todoList]);
-    console.log('indexNum= ' + indexNum)
-    console.log('todo list now equals= ' + todoList);
-  };
+      </div>
+    ))
+  );
+
   useEffect(() => {
     mapTodoList();
   }, [todoList]);
-
-  console.log(todoList);
 
   return (
     <div className="activity-column">
       <div className="column-header">
         To Do
       </div>
+
       <div>
       {mapTodoList()}
       </div>
+
     </div>
   );
 }
